Android: use the headers to have the audio pause/resume definitions and make SDL_PriveAudioData more private
