+++ src/audio/android/SDL_androidaudio.h
@@ -23,15 +23,16 @@
 #ifndef SDL_androidaudio_h_
 #define SDL_androidaudio_h_
 
-#include "../SDL_sysaudio.h"
-
-struct SDL_PrivateAudioData
-{
-    /* Resume device if it was paused automatically */
-    int resume;
-};
+#ifdef SDL_AUDIO_DRIVER_ANDROID
 
 void ANDROIDAUDIO_ResumeDevices(void);
 void ANDROIDAUDIO_PauseDevices(void);
+
+#else
+
+static void ANDROIDAUDIO_ResumeDevices(void) {}
+static void ANDROIDAUDIO_PauseDevices(void) {}
+
+#endif
 
 #endif /* SDL_androidaudio_h_ */
