+++ src/video/android/SDL_androidevents.c
@@ -28,38 +28,10 @@
 #include "../SDL_sysvideo.h"
 #include "../../events/SDL_events_c.h"
 
-/* Can't include sysaudio "../../audio/android/SDL_androidaudio.h"
- * because of THIS redefinition */
 
-#if !defined(SDL_AUDIO_DISABLED) && defined(SDL_AUDIO_DRIVER_ANDROID)
-extern void ANDROIDAUDIO_ResumeDevices(void);
-extern void ANDROIDAUDIO_PauseDevices(void);
-#else
-static void ANDROIDAUDIO_ResumeDevices(void) {}
-static void ANDROIDAUDIO_PauseDevices(void) {}
-#endif
-
-#if !defined(SDL_AUDIO_DISABLED) && defined(SDL_AUDIO_DRIVER_OPENSLES)
-extern void openslES_ResumeDevices(void);
-extern void openslES_PauseDevices(void);
-#else
-static void openslES_ResumeDevices(void)
-{
-}
-static void openslES_PauseDevices(void) {}
-#endif
-
-#if !defined(SDL_AUDIO_DISABLED) && defined(SDL_AUDIO_DRIVER_AAUDIO)
-extern void aaudio_ResumeDevices(void);
-extern void aaudio_PauseDevices(void);
-SDL_bool aaudio_DetectBrokenPlayState(void);
-#else
-static void aaudio_ResumeDevices(void)
-{
-}
-static void aaudio_PauseDevices(void) {}
-static SDL_bool aaudio_DetectBrokenPlayState(void) { return SDL_FALSE; }
-#endif
+#include "../../audio/android/SDL_androidaudio.h"
+#include "../../audio/aaudio/SDL_aaudio.h"
+#include "../../audio/openslES/SDL_openslES.h"
 
 /* Number of 'type' events in the event queue */
 static int SDL_NumberOfEvents(Uint32 type)
